User a subquery in getChildrenNodes method
The old logic In getChildrenNodes method performs a query on nodes table
 with the passed in identifier. i.e. parent identifier.

For each row that query returns we do another query in nodes, property
and addl_property table for the identifier associated in that row.
That's three sql queries per child.

In this commit we do just one subquery that joins nodes and properties
table with an outer join on addl_props. Note that nodes and properties
 have a one to one relationship but no so addl_props, a very sparse table,
 that only contains extra properties for some identifiers.

This subquery should work in MySQL 5.5 and newer releases but a more efficient subquery
using WITH will be add when we upgrade to MySQL 10.2+