getChildrenNodes method with subquery and template XML
This is another iteration to the getChildrenNodes method, where we drop
the VTD XML dom generation in favor of a simple XML template string.
I'm basically porting the existing createNode method logic, which
creates an XML Node object and populates it with its properties,
accepts, provides, etc,  and converts to a XML String which is what
returns to the getChildrenNodes method.
So the new logic basically does the same except that instead of
building an XML Dom object and then convert it to string, I simply
skip the firt part and directly use an XML String as template for
the representation of the "Node" object in XML.