Improve database SQL querying and more efficent indexing
We added multi-key indices that should work well for range queries such
as the ones with like on the identifier column. To this queries we added
the owner as a search parameter in the SQL where clause.

Two new indices were created, in both, the first key is owner
E.g.
.- nod_own_id_idx with keys on owner and identifer
.- nod_own_dep_id_idx with keys on owner, depth and identifier

Note that we've limit the size for onwer to 25 chars and for the
identifier to 250. This makes the index significantly more compact.

The new indices can be created with the vospace_update.sql
cat vospace_update | mysql -h <host> -u<user> -p vospace_prd

Note: Potentially we could drop indices: nod_own_idx (now include in the
new ones), nod_dep_idx and nod_typ_idx, both with very low cardinality
no usage for it as far as I can tell.