Improve XML serialization and container children collection
We've improved performace in two ways:
- SQL improvements:
  In the getChildrenNodes, a method dedicated to get a containers
  children, a subquery joining several tables has been used instead of
  several, independent, queries to the database for each child found under
  that container. This method is pretty ubiquous and triggered everytime a
  container is found.

- XML serialization:
  The code uses a VTD DOM, which in itself is pretty efficient but a bit
  too much for what we need, specially when the XML is very large, for
  instance, containing many thousand of children nodes.
  Now, instead of using a Node object that has an associated VTD DOM,
  we used the CADC approach of having a Node object that is DOM
  agnostic and we use this Node object, right till the end, just before
  sending it to the client, when is serialized to XML using JDOM2.

We've fully applied this DOM agnostic Node object approach only to the
"/nodes" API, however  we left in the code the origina VTD DOM logic,
except for the "getChildrenNodes" method  where we have a hybrid.
The method internally produces a JDOM2 XML that is compatible with the
original VTD format. We achieve this via the VTDXMLNodeWriter.java that
ensures it serializes it as the VTD expects it.
For the methods that fully implement the JDOM2 logic we suffix them with
the "JDOM2" string.
E.g.
  getChildrenNodesJDOM2
  getDataJDOM2
  createNodeJDOM2
  addViewsAndCapabilitiesJDOM2
  getNodeJDOM2
  etc.

The "/nodes" API is implemented in the NodeResource.java. The API hasn't
changed but now the "getNode" method in it return an
edu.noirlab.datalab.vos.Node object instead of the original
edu.caltech.vao.vospace.xml.Node one.
These objects are serialized into XML via the
edu.noirlab.datalab.vos.NodeMessageBodyWriter and the
edu.caltech.vao.vospace.NodeMessageBodyWriter respectively.