Improving transfers and results tables SQL querying
A couple of changes were made in order to make the transfer queries more
efficient so the "transfers" and "results" table can grow indefinetely
without incurring in a performance hit.

1.- A typical entry in the transfers.endpoint column looks like:
E.g.
 http://localhost:8080/vospace/vospace/data/97f13785-4fe4-4d4a-9022-6d16f045a6d6

The issue here is that a lot of searches are done using just the hash end,
 via a " like " in the where clause as:
E.g.
select <blah> from transfers where endpoint like "%97f13785-4fe4-4d4a-9022-6d16f045a6d6";

That basically means a table scan on that column, since you have to traverse
 the whole column content to get to the hash part.

An alternative to this is to simply store the contents in reverse order,
 instead of the above it'll be
6d6a540f61d6-2209-a4d4-4ef4-58731f79/atad/ecapsov/ecapsov/0808:tsohlacol//:ptth
and the where clause like query will look like:
E.g.
where rendpoint like "6d6a540f61d6-2209-a4d4-4ef4-58731f79%";

So the sql engine can quickly discriminate in the binary index.

This is already 3 to 4 times faster than the straight version with an index on
a not big table.

2.- We added an index in the transfers jobid column, since we do a join between
the transfers.jobid and the results.identifier columns.