Added more logging to VOSpace code
I tried to catch most of the exceptions and to produce some log output,
including stack trace.

The default level is INFO, which is somewhat verbose but after a period
of running and if things are OK, we can bump it to WARN.

For those processes/threads that wait for a condition to be met I added
logic so every 10 mins it sends a WARN message like the below:
E.g.
14:33:16,643  WARN 1631159213585A:TransferJob:274 - JobID:[1631159213585A] has been running for 1060 minute

If the condition take a long time to meet this message might be
perfectly fine. However there is some situations where the condition
will never be met. For instance if the client triggers a job and never
completes it. One of the reasons might be a Reset by Peer error, which
currently the client code doesn't handle but it could be other similar
reasons too.

For the above issue we would got to the database and search on the
"transfers" and "results" table. If we assert the issue should have been
completed and didn't then we just update the "transfers" table to
completed:
E.g.
mysql> select completed from transfers where jobid = '1631159213585A';
+-----------+
| completed |
+-----------+
| NULL      |
+-----------+

Then update it:
E.g.
update transfers set completed = now() where jobid = '1631159213585A';

A hanging thread can also be seen on the JVM thread stack dump:
E.g.

jstack -l $tomcat_pid | grep -E "^\"[0-9]" | sort
"1631159213585A" #59080 daemon prio=5 os_prio=31 tid=0x00007fa75709f800 nid=0x8d2f waiting on condition [0x0000700012511000]

Once the update in the DB has been performed this thread should go away.