<jobid>/results/transferDetails infinite loop
There seems to be a race condition between the client and the transfer
logic where by if the transfer job logic fails to save the details to
the "results" table, the client "storemanager" will keep pounding the
transfer "<jobid>/results/transferDetails API forever.

In addition to this the API itself can get into an infinite loop if the
phase is stuck on "EXECUTING" and no job "details" are retrieve.

I change the code to do two things:
1.- Not to get into an infinite loop, on the client side, by returning
    a default detail string if an exception occurs.
2.- Added logging so if we do get into an infinite loop, internally, we
    log this in the vospace.log file as a WARN. Also I slowed down the
    looping from a 10th of a second to a full second.