Migrate the Protocol de/serialization to JDOM2
For some extrange reason, under load, when doing an xpath on a Protocol
XML, with the VTD engine, it'll throw an Exception where the Exception
itsef was null. I don't know how that is even possible but it is
certainly weird.

Since the issue seemed to have originated in the VTD engine, I migrated
the Protocol logic to JDMO2. Basically I duplicated the Protocol class
under edu.noirlab.datalab.xml package and made everybody else point to
that.

Performance wise the JDOM2 has an initialization penalty but once it is
running it runs as fast or even faster than the VTD, so we should not
see a performance hit there.